+++ tests/generic_config_updater/multiasic_change_applier_test.py
@@ -38,7 +38,7 @@ class TestMultiAsicChangeApplier(unittest.TestCase):
             except Exception as e:
                 assert(result == False)
 
-    @patch('generic_config_updater.change_applier._get_running_config', autospec=True)
+    @patch('generic_config_updater.change_applier.ChangeApplier._get_running_config', autospec=True)
     @patch('generic_config_updater.change_applier.ConfigDBConnector', autospec=True)
     def test_apply_change_default_scope(self, mock_ConfigDBConnector, mock_get_running_config):
         # Setup mock for ConfigDBConnector
@@ -79,7 +79,7 @@ class TestMultiAsicChangeApplier(unittest.TestCase):
         # Assert ConfigDBConnector called with the correct namespace
         mock_ConfigDBConnector.assert_called_once_with(use_unix_socket_path=True, namespace="")
 
-    @patch('generic_config_updater.change_applier._get_running_config', autospec=True)
+    @patch('generic_config_updater.change_applier.ChangeApplier._get_running_config', autospec=True)
     @patch('generic_config_updater.change_applier.ConfigDBConnector', autospec=True)
     def test_apply_change_given_scope(self, mock_ConfigDBConnector, mock_get_running_config):
         # Setup mock for ConfigDBConnector
@@ -120,7 +120,7 @@ class TestMultiAsicChangeApplier(unittest.TestCase):
         # Assert ConfigDBConnector called with the correct scope
         mock_ConfigDBConnector.assert_called_once_with(use_unix_socket_path=True, namespace="asic0")
 
-    @patch('generic_config_updater.change_applier._get_running_config', autospec=True)
+    @patch('generic_config_updater.change_applier.ChangeApplier._get_running_config', autospec=True)
     @patch('generic_config_updater.change_applier.ConfigDBConnector', autospec=True)
     def test_apply_change_failure(self, mock_ConfigDBConnector, mock_get_running_config):
         # Setup mock for ConfigDBConnector
@@ -142,7 +142,7 @@ class TestMultiAsicChangeApplier(unittest.TestCase):
 
         self.assertTrue('Failed to get running config' in str(context.exception))
 
-    @patch('generic_config_updater.change_applier._get_running_config', autospec=True)
+    @patch('generic_config_updater.change_applier.ChangeApplier._get_running_config', autospec=True)
     @patch('generic_config_updater.change_applier.ConfigDBConnector', autospec=True)
     def test_apply_patch_with_empty_tables_failure(self, mock_ConfigDBConnector, mock_get_running_config):
         # Setup mock for ConfigDBConnector
